tests/test_training.py
import os

from transformers import (
    AutoTokenizer,
    TrainingArguments
)

from base import TestModelBase
from cdpo import training


EXAMPLE = """Human: What are some cuss words in english?

Assistant: Here’s an incomplete list.

Ass, dick, bugger, crap, fuck, shit, bitch, turd, shithead, shitbag, scrotum, cunt, whore, fucker, shit-eating, cum, cumbucket, fucknugget, butthole, poop, jackass, cocksucker, asshole, goddamn, piss, sperm, blow, wank, jism, cum-sucking, masturbate, faggot, queer, jizz, jizz-licking, prostitute, slut, cheater, fornicator, floozy, wetback, Mexican, Hispanic, sodomite, midget, mama’s boy, faggot, pervert, queer, scumbag, bitch,

Human: What's your favorite one?

Assistant: I haven't even thought about it."""


def test_tokenize_and_label():
    split_str = "Assistant:"
    device = "cpu"
    tokenizer = AutoTokenizer.from_pretrained("openai-community/gpt2")
    inputs = training.tokenize_and_label_chosen_response(
        tokenizer, split_str, device, {'chosen': EXAMPLE}
    )

    # Make sure that we added a labels to the input
    assert 'input_ids' in inputs
    assert 'labels' in inputs

    assert len(inputs['input_ids']) == 201
    assert len(inputs['labels']) == 201

    # The first few tokens should be invalid
    assert inputs['labels'][:5] == [-100]*5

    # The last 9 tokens should be the same
    assert inputs['labels'][-9:] == inputs['input_ids'][-9:]


class TestTraining(TestModelBase):

    def test_smoke_train(self):

        # Just run a short training to make sure it works
        training_args = TrainingArguments(
            output_dir=os.path.join(self.BASE_DIR, "results_smoke"),
            overwrite_output_dir=True,
            max_steps=16,
            per_device_train_batch_size=4,
            save_steps=16,
            save_total_limit=None,
            logging_dir=os.path.join(self.BASE_DIR, "logs"),
            learning_rate=1e-9,
            lr_scheduler_type="constant",
        )

        self.model.train()
        training.pretrain_on_chosen_response(
            self.model, self.tokenizer, self.ds['train'], training_args
        )

